GUI Configuration
Finished off creating the `Setup` and `About` windows for the GUI using `menu strip` to go at the bottom of the JukeBox Form. Also, have set the code to display the windows in runtime to change to the differentl windows smoothly in the application. In addition, tracks have been decided what will be used in the arrays.