Setting up working directory and creating/completing the dynamic array to parse the media.txt file
Parsing files with an unlimited amount of things by setting up a dynamic array from the text file media. This is done by making a list of a struct (which is a class) named GenreContents (not a list of a list) to import media into the Jukebox at the source code level. By making a struct I have simplified the code than I would have if I made a list inside of a list. For example, number 1: its name is `setup` and it has a track called `calling setup`. Essentially, this will make the array bigger b/c it’s dynamic, meaning you can add tracks to it using media txt file created in notepad. So, making a dynamic array of genres (or titles) and inside those genres will be a list of tracks inside those genres; making a list of a struct in an array thereby making a 2d Array. Furthermore, had to initialise the function “ReadMediaFile();” in public for the dynamic array to work. In addition, made 3 genres inside the dynamic array which are “general” (all the tracks I’m going to add into the MyJukeBox, plus Dr Peter O’Neill’s `Introduction.wma` track. For, “setup” title “calling setup.wma” track audited by Dr Peter O’Neill and “Hints and Tips” title for Dr Peter O’Neill’s mp3 tracks. Moreover, “iLineCounter” was included so I wouldn’t get confused meaning it will increment the array so I can see where I am in the media.txt file from the source code. For example, count 1 = the word “general” and you would ++ after the iLineCounter to increment it every time it has been used. This can be displayed by making break points whilst in “runtime” to see a clear depiction of what is occurring.  Again, right at the top of the source code I made a struct called GenreContents containing a string called name and a list of strings that are the tracks which will make each genre to hold a track where the number of tracks I input could be infinite technically. For this reason, the GenreContents struct was placed into the necessary `for loops` (needed for a dynamic array) in order to complete the dynamic array. Therefore, I have made 3 lists dynamically at the time of writing this (lists of a struct). Equally important, a setup working directory by double up ../../ which was created  in debug to build a path to the media txt file, without it the media txt file could not be located.