PopulateTextBox Function
Have began to bring text data out of the dynamic array by making 2 "for loops" and using Enviroment.Newline to bring the genres and tracks out on to seperate lines of the boxes. In addtion, multiLine must be set to "True" to display like this for a text box. Otherwise you don't need to mess around in properties for a ListBox to do this.