Simplified code and Debugged
Deleted some code and just relied on the form instead by initialling a timer and making a function for it. It was here where I made the timer to auto reset every 3 seconds in the background so when a song has finished another song will leave the playlist list box and play; the timer is always refreshing in the background and this is where it knows when to pull out another song form the playlist box. Thus, I manage to play 1 track consecutively after the other without more than one track being played at the same time. Didn’t need the Boolean flag as I used the internal playing state of the media player to play the tracks. However, the timer isn’t on the same thread with all the other variables on the user interface so can’t change the text but the music. I wasn’t able to find a come around to track this.  Although various debug techniques with Console.WriteLine was placed into my timer function in order to know if it was executing correctly, this enabled me to find the right path of the file so I could successfully play consecutive songs after the other.